refactor source strategies into separate module as per review suggestion
