+++ confidential_backend/config.py
@@ -23,7 +23,7 @@ PREFERRED_URL_SCHEME = os.getenv("PREFERRED_URL_SCHEME", 'http')
 SECONDARY_SOURCE_STRATEGIES = [
     {
         "name": "app",
-        "class": "confidential_backend.secondary_fhir_strategy.SecondaryFhirStrategy",
+        "class": "confidential_backend.source_strategies.secondary_fhir_strategy.SecondaryFhirStrategy",
         "launch_mrn_systems": os.getenv("LAUNCH_FHIR_MRN_SYSTEMS","").split(","),
         "mrn_system": os.getenv("APP_FHIR_MRN_SYSTEM"),
         "server_url": os.getenv("APP_FHIR_URL"),
